requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-APScheduler==1.13.1
gunicorn==22.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
